BOAC-3542: Adds advisor index view, refactors advisor autocomplete
