+++ scripts/db/migrate/2020/20200427-BOAC-3540/pre_deploy_01_create_note_appointment_advisor_index.sql
@@ -1,0 +1,16 @@
+BEGIN;
+
+CREATE MATERIALIZED VIEW advisor_author_index AS (
+  SELECT DISTINCT aa.* FROM (
+    SELECT a.advisor_name, a.advisor_uid
+    FROM appointments a
+    UNION
+    SELECT n.author_name AS advisor_name, n.author_uid AS advisor_uid
+    FROM notes n
+  ) aa
+  ORDER BY aa.advisor_name
+);
+  
+CREATE INDEX idx_advisor_author_index ON advisor_author_index USING btree(advisor_name);
+
+COMMIT;
