+++ src/api/notes.ts
@@ -86,20 +86,6 @@ export function deleteNote(noteId: number) {
     .then(response => response.data);
 }
 
-let $_findAuthorsByNameCancel = axios.CancelToken.source();
-
-export function findAuthorsByName(query: string, limit: number) {
-  if ($_findAuthorsByNameCancel) {
-     $_findAuthorsByNameCancel.cancel();
-  }
-  $_findAuthorsByNameCancel = axios.CancelToken.source();
-  return axios
-    .get(
-      `${utils.apiBaseUrl()}/api/notes/authors/find_by_name?q=${query}&limit=${limit}`,
-      {cancelToken: $_findAuthorsByNameCancel.token}
-    ).then(response => response.data);
-}
-
 export function addAttachment(noteId: number, attachment: any) {
   const data = {
     'attachment[0]': attachment,
